Refactor: Update Jupiter SDK
Update the Jupiter SDK to use the latest API changes,
including changes to the amount type in QuoteGet and
refactoring of several model structs.  Also update the
module import path.

+++ api_default.go
@@ -248,7 +248,7 @@ type DefaultApiQuoteGetOpts struct {
 	PreferLiquidDexes             optional.Bool
 }
 
-func (a *DefaultApiService) QuoteGet(ctx context.Context, inputMint string, outputMint string, amount int32, localVarOptionals *DefaultApiQuoteGetOpts) (QuoteResponse, *http.Response, error) {
+func (a *DefaultApiService) QuoteGet(ctx context.Context, inputMint string, outputMint string, amount uint64, localVarOptionals *DefaultApiQuoteGetOpts) (QuoteResponse, *http.Response, error) {
 	var (
 		localVarHttpMethod  = strings.ToUpper("Get")
 		localVarPostBody    interface{}
